Fixed an uninitialized variable and a few warnings.

+++ engine.cpp
@@ -163,7 +163,7 @@ void Engine::destroy_flagged(void) {
 	bool some_got_deleted;
 	do {
 		some_got_deleted = false;
-		for(int i=0; i < entities.size(); ++i) {
+		for(size_t i=0; i < entities.size(); ++i) {
 			Entity *entity = entities[i];
 			if (entity->isKilled()) {
 				some_got_deleted = true;
